Updates to drawing meshes
Not really sure if I am going to keep these changes from today or revert to a global world
and just redo the parts that need to be redone. I don't really know how much this will
affect things and I'm not sure if I should just go back and rewrite the things that are
broken with the old system (since I feel like I might be blaming this on the ECS when in
reality it's just my shitty systems that are causing the issue) or keep things how they
are (this design could be better in many ways, but I am just not sure yet and I don't
know of anyone who has used this type of engine architecture before). Ultimately, I think
it might be something between one of those two, but they are so close in terms of how
they feel to go with that I do not know if I can really make a decision about these kinds
of things. Really, I just want to make a game and have fun with it, not make stupid choices
of little actual consequince, but I guess that it the way life is sometimes. Well, I will
probably just keep this going since I would just be remaking this but with a more ECS like
model. Anyways, have fun.